Move package rulesmeta into lint
