+++ pkg/lint/rules.go
@@ -1,4 +1,4 @@
-package rulesmeta
+package lint
 
 import (
 	"github.com/santhosh-tekuri/jsonschema/v5"
