+++ pkg/lint/rules/testdata/8_missing_descriptions.json
