+++ pkg/lint/rules/testdata/8_missing_titles_and_descriptions.json
