+++ pkg/lint/rules/testdata/has_descriptions.json
