+++ pkg/lint/rules/testdata/has_titles.json
