+++ pkg/lint/rules/testdata/title_equals_description.json
@@ -1,0 +1,11 @@
+{
+    "$id": "https://example.com/person.schema.json",
+    "properties": {
+        "age": {
+            "description": "Age",
+            "title": "Age",
+            "type": "integer"
+        }
+    },
+    "type": "object"
+}
