+++ pkg/lint/rulesets/rulesets.go
@@ -5,12 +5,12 @@ import (
 
 	"github.com/santhosh-tekuri/jsonschema/v5"
 
-	"github.com/giantswarm/schemalint/pkg/lint/rulesmeta"
-	"github.com/giantswarm/schemalint/pkg/lint/rulesmeta/rules"
+	"github.com/giantswarm/schemalint/pkg/lint"
+	"github.com/giantswarm/schemalint/pkg/lint/rules"
 )
 
 type RuleSet struct {
-	rules []rulesmeta.Rule
+	rules []lint.Rule
 }
 type RuleSetName string
 
@@ -19,7 +19,7 @@ const (
 )
 
 var ClusterApp = &RuleSet{
-	rules: []rulesmeta.Rule{
+	rules: []lint.Rule{
 		rules.TitleExists{},
 		rules.DescriptionExists{},
 	},
@@ -76,10 +76,10 @@ func VerifyRuleSet(name string, schema *jsonschema.Schema) (errors []string, rec
 	for _, rule := range ruleSet.rules {
 		violations := rule.Verify(schema)
 		severity := rule.GetSeverity()
-		if severity == rulesmeta.SeverityError {
+		if severity == lint.SeverityError {
 			errors = append(errors, violations...)
 		}
-		if severity == rulesmeta.SeverityRecomendation {
+		if severity == lint.SeverityRecomendation {
 			recommendations = append(recommendations, violations...)
 		}
 	}
